+++ packages/components/src/Button/Button.tsx
@@ -1,15 +1,48 @@
 import React from "react";
+import classnames from "classnames";
+import { XOR } from "ts-xor";
+import { Icon, IconType } from "../Icon";
 
 import styles from "./Button.css";
 
-interface ButtonProps {
-  readonly label: string;
+interface ButtonBaseProps {
+  readonly type?: "primary" | "outline" | "ghost";
   onClick(): void;
 }
 
-export function Button({ label = "Button", onClick }: ButtonProps) {
+interface ButtonIconProps extends ButtonBaseProps {
+  readonly icon: IconType;
+  /**
+   * The label that is to be read by screen readers.
+   */
+  readonly ariaLabel: string;
+  readonly label?: string;
+}
+
+interface ButtonNoIconProps extends ButtonBaseProps {
+  readonly label: string;
+  readonly icon?: IconType;
+}
+
+type ButtonProps = XOR<ButtonIconProps, ButtonNoIconProps>;
+
+export function Button({
+  label,
+  icon,
+  type = "primary",
+  ariaLabel,
+  onClick,
+}: ButtonProps) {
+  const buttonClasses = classnames(styles.button, styles[type]);
+
   return (
-    <button type="button" onClick={handleClick} className={styles.button}>
+    <button
+      type="button"
+      onClick={handleClick}
+      className={buttonClasses}
+      aria-label={ariaLabel}
+    >
+      {icon && <Icon icon={icon} />}
       {label}
     </button>
   );