+++ packages/components/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.6.0](https://github.com/eddysims/pids/compare/@pids/components@0.5.1...@pids/components@0.6.0) (2021-02-25)
+
+
+### Features
+
+* add type and icon props to Button ([#14](https://github.com/eddysims/pids/issues/14)) ([85c1efb](https://github.com/eddysims/pids/commit/85c1efbd5f43a65f108324ab7388afb1e8e3e678))
+
+
+
+
+
+## [0.5.1](https://github.com/eddysims/pids/compare/@pids/components@0.5.0...@pids/components@0.5.1) (2021-02-24)
+
+**Note:** Version bump only for package @pids/components
+
+
+
+
+
+# [0.5.0](https://github.com/eddysims/pids/compare/@pids/components@0.4.0...@pids/components@0.5.0) (2021-02-24)
+
+
+### Features
+
+* adds the Text component ([#16](https://github.com/eddysims/pids/issues/16)) ([88b5b63](https://github.com/eddysims/pids/commit/88b5b63da8ad635613a57751bdb260ad5ef10baa))
+
+
+
+
+
+# [0.4.0](https://github.com/eddysims/pids/compare/@pids/components@0.3.0...@pids/components@0.4.0) (2021-02-24)
+
+
+### Features
+
+* add styling to Heading ([#15](https://github.com/eddysims/pids/issues/15)) ([a69e0ff](https://github.com/eddysims/pids/commit/a69e0ff0cc41fed3c1d4a93e00222ea41c046e95))
+
+
+
+
+
 # [0.3.0](https://github.com/eddysims/pids/compare/@pids/components@0.2.0...@pids/components@0.3.0) (2021-02-21)